implementing auth , notification routes

+++ routes/notification_routes.py
@@ -1,13 +1,21 @@
 from flask import Blueprint, request, jsonify
 from extensions import db
 from models.notification import Notification
+#Implementing Auth
+from utils.auth_helpers import role_required
+from flask_jwt_extended import get_jwt_identity
+
 
 notification_bp = Blueprint('notifications', __name__)
+
+
+# Auth logic :Create: admin or system only.        View/Mark Read/Delete: A user should only be able to interact with their own notifications. We use get_jwt_identity() to enforce this.
 
  # OPTIMIZATION:Implementing pagination to prevent memory crashes
 
 # --- CREATE: Generate a new notification ---
 @notification_bp.route('/', methods=['POST'], strict_slashes=False)
+@role_required('admin')
 def create_notification():
     data = request.get_json()
     
@@ -31,7 +39,13 @@ def create_notification():
 
 # --- READ: Get paginated notifications for a specific user ---
 @notification_bp.route('/user/<int:user_id>', methods=['GET'], strict_slashes=False)
+@role_required('admin', 'zone_operator', 'resident')
 def get_user_notifications(user_id):
+
+    # SECURITY CHECK: Ensuring the requester is asking for their own data
+    current_user_id = int(get_jwt_identity())
+    if current_user_id != user_id:
+        return jsonify({"error": "Unauthorized to view these notifications"}), 403
    
     page = request.args.get('page', 1, type=int)
     per_page = request.args.get('per_page', 20, type=int)
@@ -50,8 +64,14 @@ def get_user_notifications(user_id):
 
 # --- UPDATE: Mark a notification as read ---
 @notification_bp.route('/<int:id>/read', methods=['PUT'], strict_slashes=False)
+@role_required('admin', 'zone_operator', 'resident')
 def mark_as_read(id):
     notification = Notification.query.get_or_404(id)
+
+    # SECURITY CHECK
+    current_user_id = int(get_jwt_identity())
+    if notification.user_id != current_user_id:
+        return jsonify({"error": "Unauthorized"}), 403
     
     notification.is_read = True
     db.session.commit()
@@ -60,8 +80,14 @@ def mark_as_read(id):
 
 # --- DELETE: Remove a notification ---
 @notification_bp.route('/<int:id>', methods=['DELETE'], strict_slashes=False)
+@role_required('admin', 'zone_operator', 'resident')
 def delete_notification(id):
     notification = Notification.query.get_or_404(id)
+
+    # SECURITY CHECK
+    current_user_id = int(get_jwt_identity())
+    if notification.user_id != current_user_id:
+        return jsonify({"error": "Unauthorized"}), 403
     
     db.session.delete(notification)
     db.session.commit()
